ADDTOWALLET functionality working - includes addition of context to app.js. Separated games data into context folder.
