+++ PostgreSQL/schema.sql
@@ -23,7 +23,7 @@ CREATE TABLE users (
   wallet_address VARCHAR NULL DEFAULT NULL
 );
 COPY users(id, username, tokens, phone_number, wallet_address)
-FROM '/Users/tim/Documents/HackReactor/Blue-Ocean-Ravenclaw/fakeData/user.csv'
+FROM '/Users/austinchapin/Desktop/HackReactor/Senior Phase/BlueOcean/luckylucy/fakeData/user.csv'
 DELIMITER ','
 CSV HEADER;
 
@@ -46,7 +46,7 @@ CREATE TABLE nfts (
 
 COPY nfts(id, id_user, description, external_url, image, name, value)
 
-FROM '/Users/tim/Documents/HackReactor/Blue-Ocean-Ravenclaw/fakeData/nfts.csv'
+FROM '/Users/austinchapin/Desktop/HackReactor/Senior Phase/BlueOcean/luckylucy/fakeData/nfts.csv'
 DELIMITER ','
 CSV HEADER;
 
@@ -64,7 +64,7 @@ CREATE TABLE card_inventory (
   quantity INTEGER NULL DEFAULT NULL
 );
 COPY card_inventory(id, id_user, card_name, quantity)
-FROM '/Users/tim/Documents/HackReactor/Blue-Ocean-Ravenclaw/fakeData/card_inventory.csv'
+FROM '/Users/austinchapin/Desktop/HackReactor/Senior Phase/BlueOcean/luckylucy/fakeData/card_inventory.csv'
 DELIMITER ','
 CSV HEADER;
