+++ pages/api/cards/[username].js
@@ -9,7 +9,7 @@ export default async function handler(req, res) {
       UPDATE card_inventory
       SET quantity = quantity + $1
       WHERE id_user = (SELECT id FROM users WHERE username = $2)
-      AND WHERE card_name = $3;
+      AND card_name = $3;
     `;
     query.values = [req.body.quantity, req.query.username, req.body.card_name];
