+++ pages/games/index.jsx
@@ -1,5 +1,8 @@
 import * as React from 'react';
 import { useState, useEffect, useContext } from 'react';
+import { useAppContext } from '../../context/state.js';
+import axios from 'axios';
+
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
@@ -9,11 +12,6 @@ import CardContent from '@mui/material/CardContent';
 import CardHeader from '@mui/material/CardHeader';
 import CssBaseline from '@mui/material/CssBaseline';
 import Grid from '@mui/material/Grid';
-import CardMedia from '@mui/material/CardMedia';
-import StarIcon from '@mui/icons-material/StarBorder';
-import CasinoIcon from '@mui/icons-material/Casino';
-import Filter3Icon from '@mui/icons-material/Filter3';
-import Filter7Icon from '@mui/icons-material/Filter7';
 import AddIcon from '@mui/icons-material/Add';
 import RemoveIcon from '@mui/icons-material/Remove';
 import IconButton from '@mui/material/IconButton';
@@ -23,7 +21,8 @@ import Link from '@mui/material/Link';
 import GlobalStyles from '@mui/material/GlobalStyles';
 import Container from '@mui/material/Container';
 import Modal from '@mui/material/Modal';
-import AppContext from '../_app.jsx';
+import { games } from '../../context/games.js';
+
 
 const style = {
   position: 'absolute',
@@ -53,65 +52,17 @@ const style = {
 };
 
 
-const games = [
-  {
-    title: "High Roller",
-    price: 10,
-    description: [
-      'Spin 3 dice',
-      'and win a',
-      'CUSTOM NFT',
-      'Ready to roll?',
-    ],
-    buttonText: 'BUY NOW',
-    buttonVariant: 'contained',
-    route: '/games/dice',
-    icon: <CasinoIcon />,
-  },
-  {
-    title: 'BINGO',
-    subheader: 'Most popular',
-    price: 20,
-    description: [
-      'Get 3 BINGO boards,',
-      'see if your numbers',
-      'are the lucky ones!',
-      'WIN BIG TODAY!',
-    ],
-    buttonText: 'BUY NOW',
-    buttonVariant: 'contained',
-    route: '/games/bingo',
-    icon: <Filter3Icon />
-  },
-  {
-    title: "Lucky 7's",
-    price: 15,
-    description: [
-      'Coming soon to',
-      'cRyPtOcAsInO',
-      'Scratch off to reveal',
-      'your big win!'
-    ],
-    buttonText: 'COMING SOON...',
-    buttonVariant: 'contained',
-    route: '/games/lucky7',
-    icon: <Filter7Icon/>
-  },
-];
-
-
-
-
 
 function PricingContent() {
-
-  const pickUpContext = useContext(AppContext)
 
   const [open, setOpen] = useState(false);
   const [gameCount, setGameCount] = useState(1);
   const [gameTitle, setGameTitle] = useState('');
   const [total, setTotal] = useState(0);
   const [game, setGame] = useState({});
+
+  const context = useAppContext();
+  const {tokens} = useAppContext();
 
 
   const handleOpen = (e) => {
@@ -148,11 +99,19 @@ function PricingContent() {
   }
 
   const handleAddToWallet = () => {
-    // Need to deduct points from wallet
-    // Need to add gameCount of game Title to database
-    gameCount // # of games to add
-    game.title // String of game title
 
+    if (total > tokens) {
+      console.log('BROKE')
+    } else {
+      axios.post(`/api/tokens/${context.username}`, { tokens: (total * -1) })
+        .then((res) => {
+          console.log(res);
+          axios.put(`/api/cards/${context.username}`, { card_name: game.dbTitle, quantity: gameCount })
+            .then((res) => console.log(res))
+            .catch((err) => console.log(err));
+        })
+        .catch((err) => console.log(err));
+    }
     handleClose();
   }
 
@@ -163,7 +122,7 @@ function PricingContent() {
 
   return (
     <React.Fragment>
-      {console.log(pickUpContext)}
+      {console.log(context)}
       <GlobalStyles styles={{ ul: { margin: 0, padding: 0, listStyle: 'none' } }} />
       <CssBaseline />
       <AppBar
@@ -306,14 +265,3 @@ function PricingContent() {
 export default function Pricing() {
   return <PricingContent />;
 }
-
-
-
-// POST 'api/tokens/username'
-// POST 'api/cards/username'
-
-
-// {tokens: -500}
-// {card_number: highroller, quantity: 5}
-
-// GET 'api/userpage/username'
